refactor: split xctestrun product resolution
